feat: using kafka external streams as sinks.
Supports the following use cases:
* `INSERT INTO <kafka-external-stream> (<cols...>) VALUES (<values...>)`
* `INSERT INTO <kafka-external-stream> SELECT <cols...> FROM <other-stream>`
* `CREATE MATERIALIZED VIEW <view-name> INTO <kafka-external-stream> AS ...`

Messages will be encoded with the format sepcified with the
`data_format` setting of the external stream. All output formats defined
in `Formats/registerFormats.cpp` should be supported. However, not all
supported formats are tested.

A new setting `properties` is added for creating kafka external streams.
This allows users to customize the kafka client used by the kafka
external stream without us adding more settings in the future. Not all
properties are supported, currently supported properties are:
* enable.idempotence
* message.timeout.ms
* queue.buffering.max.messages
* queue.buffering.max.kbytes
* queue.buffering.max.ms
* message.send.max.retries
* retries
* retry.backoff.ms
* retry.backoff.max.ms
* batch.num.messages
* batch.size
* compression.codec
* compression.type
* compression.level

Note that, this `properties` setting is only used when the external
stream is used as a sink. We can add support for source later. Also,
currently `properties` is validated on runtime, we could improve it to
validate it when the stream is being created.

This initial implementation does not support specifying message key.

Currently there is no good way to add tests for this feature yet. I will
firstly utilize our integration test to add some tests from there. Later
on we can enhance the test framework to support testing this feature.