feature: kafka external stream partitioning (#273)
A new feature that adds a new setting name sharding_expr for creating kafka external stream. sharding_expr accepts a SQL expression which is evaluated into an integer, and the value of that expression is used to decide to which partition the message is sent. For example:

CREATE EXTERNAL STREAM foo (
a string,
b int32,
) type='kafka',brokers='localhost:9092',topic='foo',sharding_expr='to_int(b)'
The KafkaSink will use the expression to divide each in-coming Chunk into Blocks by the sharding key value and then send the Blocks to the Kafka topic.

closes #192